+++ internal/pkg/client/shub/pull.go
@@ -196,7 +196,7 @@ func Pull(ctx context.Context, imgCache *cache.Handle, pullFrom, tmpDir string, 
 }
 
 // PullToFile will pull a shub image to the specified location, through the cache, or directly if cache is disabled
-func PullToFile(ctx context.Context, imgCache *cache.Handle, pullTo, pullFrom string, noHTTPS bool) (imagePath string, err error) {
+func PullToFile(ctx context.Context, imgCache *cache.Handle, pullTo, pullFrom string, noHTTPS, sandbox bool) (imagePath string, err error) {
 	directTo := ""
 	if imgCache.IsDisabled() {
 		directTo = pullTo
@@ -208,7 +208,7 @@ func PullToFile(ctx context.Context, imgCache *cache.Handle, pullTo, pullFrom st
 		return "", fmt.Errorf("error fetching image to cache: %v", err)
 	}
 
-	if directTo == "" {
+	if directTo == "" && !sandbox {
 		// mode is before umask if pullTo doesn't exist
 		err = fs.CopyFileAtomic(src, pullTo, 0o777)
 		if err != nil {
@@ -216,5 +216,10 @@ func PullToFile(ctx context.Context, imgCache *cache.Handle, pullTo, pullFrom st
 		}
 	}
 
+	if sandbox {
+		if err := client.ConvertSifToSandbox(directTo, src, pullTo); err != nil {
+			return "", err
+		}
+	}
 	return pullTo, nil
 }