+++ CHANGELOG.md
@@ -42,6 +42,7 @@ For older changes see the [archived Singularity change log](https://github.com/a
 - Suppress unneeded warnings and remove default silent message level for the
   post script instance.
 - Add `dnf` definition file bootstrap options as an alias to `yum`.
+- Add --sandbox option to apptainer pull.
 
 ## Changes for v1.3.x